BitTorrent btpd 0.13 package
this patch also add build-in sha1 functions and remove openssl dependency

tested on current kamikaze linux 2.4 in wl500gP

About:
BTPD is a bittorrent client consisting of a daemon and client commands, which can be used to read and/or manipulate the daemon state. The daemon is capable of running several torrents simultaneously and only uses one tcp port. It's fairly low on resource usage and should be perfect for file distribution sites. Efficient downloads and ease of use makes this client a good choice for the casual user as well.